change migration add id voucher on table orders

+++ database/migrations/2020_12_22_120502_add_id_voucher_on_orders_table.php
@@ -14,7 +14,7 @@ class AddIdVoucherOnOrdersTable extends Migration
     public function up()
     {
         Schema::table('orders', function($table) {
-            $table->unsignedInteger('id_voucher')->nullable()->after('phone');
+            $table->bigInteger('id_voucher')->nullable()->after('phone');
             $table->foreign('id_voucher')->references('id')->on('vouchers');
         });
     }
